+++ spotlight/src/main/java/com/takusemba/spotlight/CustomTarget.java
@@ -5,6 +5,8 @@ import android.graphics.PointF;
 import android.support.annotation.LayoutRes;
 import android.view.View;
 
+import com.takusemba.spotlight.shapes.Shape;
+
 /**
  * Target
  *
