+++ spotlight/src/main/java/com/takusemba/spotlight/SpotlightView.java
@@ -17,6 +17,8 @@ import android.util.AttributeSet;
 import android.view.View;
 import android.widget.FrameLayout;
 
+import com.takusemba.spotlight.shapes.Shape;
+
 /**
  * Spotlight View
  *
